cookbook/google-adk/pyproject.toml
[project]
name = "simlab-google-adk"
version = "0.1.0"
description = "Google ADK agent cookbook with a SimLab BaseAgent adapter"
readme = "google-adk.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.28.0",
    "litellm<=1.82.6",
    "simulationlab",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/google_adk_agent"]

[tool.uv.sources]
simulationlab = { path = "../..", editable = true }

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["INP001"]

